+++ src/blocks/attendance-selector/edit.js
@@ -2,10 +2,33 @@
  * External dependencies.
  */
 import { useBlockProps } from '@wordpress/block-editor';
+import { Button, Modal } from "@wordpress/components";
+import { useState } from "@wordpress/element";
+
 /**
  * Internal dependencies.
  */
 import AttendanceSelector from '../../components/AttendanceSelector';
+const AttendanceModal = () => {
+	const [isOpen, setOpen] = useState(false);
+	const openModal = () => setOpen(true);
+	const closeModal = () => setOpen(false);
+
+	return (
+		<>
+			<Button variant="secondary" onClick={openModal}>
+				Open Modal
+			</Button>
+			{isOpen && (
+				<Modal title="This is my modal" onRequestClose={closeModal}>
+					<Button variant="secondary" onClick={closeModal}>
+						My custom close button
+					</Button>
+				</Modal>
+			)}
+		</>
+	);
+};
 
 const Edit = () => {
 	const blockProps = useBlockProps();
@@ -22,6 +45,7 @@ const Edit = () => {
 				currentUser={ currentUser }
 				type={ 'upcoming' }
 			/>
+			<AttendanceModal />
 		</div>
 	);
 };
