+++ src/blocks/venue-information/block.json
@@ -30,5 +30,6 @@
 	},
 	"textdomain": "gatherpress",
 	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
 	"style": "file:./style-index.css"
 }
