+++ src/blocks/venue-information/index.js
@@ -7,6 +7,7 @@ import { registerBlockType } from '@wordpress/blocks';
  * Internal dependencies.
  */
 import edit from './edit';
+import save from './save';
 
 import metadata from './block.json';
 
@@ -14,5 +15,5 @@ import './style.scss';
 
 registerBlockType( metadata, {
 	edit,
-	save: () => null,
+	save,
 } );
